fixed CUITe_BrowserWindow.Launch by using BrowserWindow.CopyFrom updated browser window tests to use generic \ non browser specific window titles added DataManager_GetDataRowUsingEmbeddedResourceFromTypeAsString_Succeeds test updated DynamicBrowserWindowTitleRepository and TestHtmlPage to use properties to declare html elements which use "this.Get<T>" instead of field initializers removed hintpath for UnitTestFramework assembly fixed Launch_ObjectRepositoryTempHtmlFile_CanFindUnorderedListsByTagAndClassName test method added SetText_OnHtmlPassword_Succeeds test fixed html5 control sample test method fixed Launch_ObjectRepositoryTempHtmlFile_CanFindUnorderedListsByTagAndClassName test fixed SelectItem_UsingHtmlComboBoxThatAlertsOnChange_Succeeds

+++ Sample_CUITeTestProject/TempFile.cs
@@ -16,15 +16,19 @@ namespace Sample_CUITeTestProject
         /// </value>
         public string FilePath { get; private set; }
 
+        public bool DeleteFileOnDispose { get; set; }
+
         /// <summary>
         /// Initializes a new instance of the <see cref="TempFile"/> class.
         /// </summary>
         /// <param name="contents">The contents.</param>
         public TempFile(string contents)
         {
-            FilePath = Path.GetTempFileName();
+            FilePath = Path.GetTempFileName() + ".html";
 
             File.WriteAllText(FilePath, contents);
+
+            this.DeleteFileOnDispose = true;
         }
 
         /// <summary>
@@ -32,7 +36,10 @@ namespace Sample_CUITeTestProject
         /// </summary>
         public void Dispose()
         {
-            File.Delete(FilePath);
+            if (this.DeleteFileOnDispose)
+            {
+                File.Delete(FilePath);
+            }
         }
     }
 }
